--- .gitmodules
@@ -1,7 +1,0 @@
-[submodule "spring/src/main/resources"]
-	path = spring/src/main/resources
-	url = https://github.com/LlamaVista/security.git
-	branch = master
-[submodule "flask/env"]
-	path = flask/env
-	url = https://github.com/LlamaVista/flask_env.git
