+++ FastAPI/database/__init__.py
@@ -1,0 +1,3 @@
+from database.models import User
+from database.mongo import MongoClient
+from database.mongo import router
