+++ FastAPI/database/models/__init__.py
@@ -1,0 +1,2 @@
+from .object import ObjectIdPydanticAnnotation
+from .user import User
