+++ FastAPI/service/__init__.py
@@ -1,0 +1,1 @@
+from .google_auth import router
