--- spring/build.gradle
@@ -1,41 +1,0 @@
-plugins {
-	id 'java'
-	id 'org.springframework.boot' version '3.1.5'
-	id 'io.spring.dependency-management' version '1.1.3'
-}
-
-group = 'hoon.capstone'
-version = '0.0.1-SNAPSHOT'
-
-java {
-	sourceCompatibility = '17'
-}
-
-configurations {
-	compileOnly {
-		extendsFrom annotationProcessor
-	}
-}
-
-repositories {
-	mavenCentral()
-}
-
-dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	implementation group: 'com.microsoft.azure', name: 'azure-spring-boot-starter', version: '2.2.0'
-	implementation group: 'com.azure', name: 'azure-storage-blob', version: '12.24.0'
-	implementation group: 'org.springframework.boot', name: 'spring-boot-starter-data-jpa', version: '3.1.1'
-	implementation group: 'mysql', name: 'mysql-connector-java', version: '8.0.33'
-	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
-	runtimeOnly 'com.microsoft.sqlserver:mssql-jdbc'
-
-	compileOnly 'org.projectlombok:lombok'
-	annotationProcessor 'org.projectlombok:lombok'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	testImplementation 'org.mockito:mockito-core'
-}
-
-tasks.named('test') {
-	useJUnitPlatform()
-}
