--- spring/src/main/java/hoon/capstone/llama/LlamaApplication.java
@@ -1,13 +1,0 @@
-package hoon.capstone.llama;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.scheduling.annotation.EnableAsync;
-
-@SpringBootApplication
-public class LlamaApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(LlamaApplication.class, args);
-	}
-}
