--- spring/src/main/java/hoon/capstone/llama/config/WebConfig.java
@@ -1,27 +1,0 @@
-package hoon.capstone.llama.config;
-
-import jakarta.annotation.Nullable;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.Objects;
-
-@Configuration
-public class WebConfig {
-
-    @Bean
-    public WebMvcConfigurer corsConfigurer() {
-        return new WebMvcConfigurer() {
-            @Override
-            public void addCorsMappings(@Nullable CorsRegistry registry) {
-                Objects.requireNonNull(registry).addMapping("/**")
-                        .allowedOrigins("http://localhost:5901", "http://localhost:8080", "http://localhost:3000")
-                        .allowedMethods("GET", "POST", "PUT", "DELETE")
-                        .allowedHeaders("*")
-                        .allowCredentials(true);
-            }
-        };
-    }
-}
