--- spring/src/main/java/hoon/capstone/llama/domain/Grammar.java
@@ -1,18 +1,0 @@
-package hoon.capstone.llama.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum Grammar {
-    SEABORN("Seaborn"),
-    ALTAIR("Altair"),
-    MATPLOTLIB("MatPlotLib"),
-    GGPLOT("GGPlot");
-
-    private final String library;
-
-    Grammar(String library) {
-        this.library = library;
-    }
-
-}
