--- spring/src/main/java/hoon/capstone/llama/domain/Result.java
@@ -1,13 +1,0 @@
-package hoon.capstone.llama.domain;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class Result {
-    private String step;
-    private Object result;
-}
