--- spring/src/main/java/hoon/capstone/llama/service/RequestService.java
@@ -1,39 +1,0 @@
-package hoon.capstone.llama.service;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import hoon.capstone.llama.domain.Settings;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-
-@Service
-public class RequestService {
-    private final RestTemplate restTemplate;
-
-    public RequestService(RestTemplate restTemplate) {
-        this.restTemplate = restTemplate;
-    }
-
-    public String sendFileAndData(MultipartFile file, Settings settings, String userMessage, String flaskUrl)
-            throws IOException {
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.MULTIPART_FORM_DATA);
-
-        MultiValueMap<String, Object> body = new LinkedMultiValueMap<>();
-        body.add("file", file.getResource());
-        body.add("settings", new ObjectMapper().writeValueAsString(settings));
-        body.add("message", userMessage);
-
-        HttpEntity<MultiValueMap<String, Object>> requestEntity = new HttpEntity<>(body, headers);
-        ResponseEntity<String> response = restTemplate.postForEntity(flaskUrl, requestEntity, String.class);
-        return response.getBody();
-    }
-}
