--- spring/src/test/java/hoon/capstone/llama/LlamaApplicationTests.java
@@ -1,13 +1,0 @@
-package hoon.capstone.llama;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LlamaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
